refactored catpanions
