+++ client/src/components/CatpanionsList.js
@@ -1,16 +1,13 @@
 import React, { useEffect, useState } from 'react'
 import Stack from "@mui/material/Stack";
-import Button from '@mui/material/Button';
-import Avatar from '@mui/material/Avatar';
-import { Grid, Paper } from "@mui/material"
-import { Link as RouterLink } from 'react-router-dom'
+import { Paper } from "@mui/material"
 import { styled } from "@mui/material/styles";
 import CatpanionItem from "./CatpanionItem.js"
+import CatpanionSearchItem from "./CatpanionSearchItem.js"
 
 function CatpanionsList( { currentUser, search, users, catpanions } ) {
 
     const [loggedInUser, setLoggedInUser] = useState({})
-    const [searchDisabled, setSearchDisabled] = useState(false)
 
     const Item = styled(Paper)(({ theme }) => ({
         ...theme.typography.body2,
@@ -26,68 +23,19 @@ function CatpanionsList( { currentUser, search, users, catpanions } ) {
     }, [catpanions])
 
     function allUsers(){
-
-        const friendIdsArray = []
-        loggedInUser.friends?.map(friend => {
-            return friendIdsArray.push(friend.id)
-        })
         
         let allUsersArray = []
         const matchingUsers = users.filter(user => {
             return user.full_name.toLowerCase().includes(search.toLowerCase())
         })
 
-        console.log(matchingUsers)
-        console.log(allUsersArray)
-
         allUsersArray = matchingUsers
 
         const renderUsers = allUsersArray?.map(user => {
-            return <Item key={user.id}>
-                        <Grid container>
-                            <Grid item>
-                                <Avatar alt="Remy Sharp" src={user.purrfile_picture} sx={{ width: "95px", height: "95px" }}/>
-                            </Grid>
-                            <Grid item sx={{ flexGrow: 1 }}>
-                                <h2>{user.full_name}</h2>
-                                <h4>@{user.username}</h4>
-                            </Grid>
-                            <Grid item marginLeft="2%">
-                                {friendIdsArray.includes(user.id) ? 
-                                <>    
-                                    <Button variant="contained" disabled={searchDisabled} sx={ { borderRadius: 28, mr: 2, backgroundColor: "#33691e" } } component={RouterLink} to="/messaging">Message</Button>
-                                    <Button variant="contained" disabled={searchDisabled} sx={ { borderRadius: 28, backgroundColor: "#33691e" } } onClick={
-                                        function handleDeleteFriend(){
-                                            const catpanionToDelete = catpanions?.find(catpanion => {
-                                                if ((catpanion.user_id === loggedInUser.id) && (catpanion.friend_id === user.id)){
-                                                    return catpanion.id
-                                                }
-                                            })
-                                            fetch(`/catpanions/${catpanionToDelete.id}`, {method: "DELETE"})
-                                            setSearchDisabled(currentState => !currentState)
-                                        }
-                                    }>{searchDisabled ? "Friend Removed" : "Unfriend"}</Button>
-                                </>
-                                :
-                                    <Button variant="contained" sx={ { borderRadius: 28 } } onClick={
-                                        function handleAddFriend(){
-                                            fetch(`/catpanions`, {
-                                                method: 'POST',
-                                                headers: {
-                                                    "Accept": "application/json",
-                                                    "Content-Type": "application/json"
-                                                },
-                                                body: JSON.stringify({user_id: currentUser.id, friend_id: user.id}),
-                                            })
-                                        }
-                                    }>Add Friend</Button>
-                            }
-                            </Grid>
-                        </Grid>
-                    </Item>
-        })
+            return <CatpanionSearchItem key={user.id} user={user} catpanions={catpanions} loggedInUser={loggedInUser} Item={Item} currentUser={currentUser}/>
+            })
         return renderUsers
-    }
+        }
 
   return (
     <Stack
